Add updateDriver and deleteDriver lengths/sizes[] from into to uint8_t
Update examples and driver: int pins[] to uint8_t pins[], int lengths[] to uint16_t lenghts

IDF5.5
- Add updateDriver and deleteDriver
- set MAX_PINS to 20 (from 16)

+++ examples/gettingstarted/gettingstarted.ino
@@ -5,9 +5,9 @@
 uint8_t leds[NUMSTRIPS*NUM_LEDS_PER_STRIP*3];
 
 #ifdef CONFIG_IDF_TARGET_ESP32S3
-int pins[6] = {9, 10,12,8,18,17};
+uint8_t pins[6] = {9, 10,12,8,18,17};
 #else
-int pins[6] = {14, 12, 13, 25, 33, 32};
+uint8_t pins[6] = {14, 12, 13, 25, 33, 32};
 #endif
 
 I2SClocklessLedDriver driver;
